+++ include/common/RowsetToSQL.h
@@ -303,8 +303,8 @@ void ClassifyColumnTypes(std::vector<ReaderColumnContentsType> &into, Reader &re
 
 template <class Reader>
 bool ReaderToSQL(const char *table_name, const char *key_field, FILE *out, Reader &reader) {
-    size_t col;
-    size_t cc = reader.ColumnCount();
+    uint32 col;
+    uint32 cc = reader.ColumnCount();
     int key_col = -1;
 
     if(key_field != NULL) {
